Start refactoring app cast generator functionality to its own class
